Have student assessment goals able to be 'uncompleted'
The 'goalStates' in LevelSessions generally does not update goals
that go from 'success' to something else, so that achievements are not
lost in this scenario. But the teacher assessments view relies on this
property to show the current student state. Have the property able
to be unset for students, since they aren't using achievements anyway.